DL/Avanzado/Vision-Transformer/requirements.txt
matplotlib==3.3.4
numpy==1.18.5
opencv-contrib-python==4.1.2.30
Pillow==8.1.2
protobuf==3.15.5
tensorflow==2.3.0
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0
tensorflow-estimator==2.3.0
tensorflow-metadata==0.28.0
tqdm==4.59.0